Client/requirements.txt
Flask==0.12.3
busio
digitalio
board
adafruit_mcp3xxx.mcp3008
adafruit_mcp3xxx.analog_in
RPi.GPIO